requirements.txt
annotated-types==0.5.0
anyio==3.7.1
audiostretchy==1.3.5
azure-cognitiveservices-speech==1.31.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
exceptiongroup==1.1.2
fastapi==0.100.0
fire==0.5.0
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-core==2.3.3
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
idna==3.4
numpy==1.24.4
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.0.2
pydantic-core==2.1.2
pydub==0.25.1
pymp3==0.1.9
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soxr==0.3.7
starlette==0.27.0
termcolor==2.4.0
typing-extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
